refactor: drop unused ArtifactViewModeContext
The context's only reader was the useArtifactAutoOpen hook removed in
a87874dd. The file shrinks to the ArtifactViewMode type + default
(renamed artifactViewMode.ts) and ChatProvider loses the dead provider
layer. Reintroduce alongside the hook if the SDK integration needs it.
