+++ packages/react-headless/src/store/ChatProvider.tsx
@@ -7,7 +7,7 @@ import {
   buildArtifactRendererRegistry,
 } from "./ArtifactRenderersContext";
 import { ArtifactStorageContext } from "./ArtifactStorageContext";
-import { ArtifactViewModeContext, DEFAULT_ARTIFACT_VIEW_MODE } from "./ArtifactViewModeContext";
+import { DEFAULT_ARTIFACT_VIEW_MODE } from "./artifactViewMode";
 import { ChatContext } from "./ChatContext";
 import { createChatStore } from "./createChatStore";
 import { createDetailedViewStore } from "./createDetailedViewStore";
@@ -81,11 +81,7 @@ export const ChatProvider: FC<ChatProviderProps> = ({
           <ArtifactRenderersContext.Provider value={artifactRendererRegistry}>
             <ArtifactStorageContext.Provider value={resolvedStorage.artifact ?? null}>
               <ArtifactCategoriesContext.Provider value={artifactCategories ?? EMPTY_CATEGORIES}>
-                <ArtifactViewModeContext.Provider
-                  value={artifactViewMode ?? DEFAULT_ARTIFACT_VIEW_MODE}
-                >
-                  {children}
-                </ArtifactViewModeContext.Provider>
+                {children}
               </ArtifactCategoriesContext.Provider>
             </ArtifactStorageContext.Provider>
           </ArtifactRenderersContext.Provider>
