+++ packages/react-headless/src/store/artifactViewMode.ts
@@ -1,0 +1,3 @@
+export type ArtifactViewMode = "auto-open" | "open-on-mount" | "overview";
+
+export const DEFAULT_ARTIFACT_VIEW_MODE: ArtifactViewMode = "overview";
